Mock the request Mock the request 1. using nock to mock request 2. separate the run tests
