+++ package.json
@@ -18,7 +18,9 @@
   },
   "devDependencies": {
     "fs-extra": "^4.0.2",
+    "nock": "^9.0.27",
     "request-debug": "^0.2.0",
-    "tape": "^4.8.0"
+    "tape": "^4.8.0",
+    "uuid": "^3.1.0"
   }
 }
